Move `get_name_for_type()` to common module
